Move the source fetchers into leads-core
Finding the conversations is the half of this product a vertical would
be built on, and it was the half still welded to the app: five source
families behind ScrapeCreators, two env vars for Reddit's personal feed
token, and a Browserless fallback reached through a dynamic import.

They move behind `createSources(deps)`. A factory rather than a deps
argument per function, because radarScan passes the searchers as
callbacks for its dynamic queries and radarEngage calls fetchRedditText
on its own — bound functions keep those call sites unchanged.

Two decisions worth naming. `redditAuth` is a function, not an object:
the original read env on every call, and freezing the credentials when
the factory is built would mean an environment change never arrives —
there is a test for it. And the Browserless page script stays in the
app: the package is handed "a function that turns a URL into text" and
does not know a browser is behind it.

Adds a test that did not exist, on routing each source kind to its
endpoint. That is the bug class the comments in this file describe:
linkedin_query fell through to fetchFeed and fetched the keywords as if
they were an RSS url, and X was called with `id=` where the endpoint
wants `url=`. Both surfaced as "0 items" — a clean empty success,
indistinguishable from a quiet day. Now that inversion fails a test
instead of silently killing a source.

radar.ts loses ~340 lines net.
