+++ changelog/2026-09-01-leads-core-package.md
@@ -6,7 +6,7 @@ hunting reggerebbero un SaaS verticale a sé? La risposta onesta era "sì, ma no
 director, scheduler ed email. Quello che ha valore fuori da Anomalia era dentro, ma non
 separabile a vista.
 
-Ora quattro moduli stanno in `packages/leads-core/`, e il guardiano che già esisteva
+Ora cinque moduli stanno in `packages/leads-core/`, e il guardiano che già esisteva
 (`packages/no-app-imports.test.ts`) li tiene puliti: al primo `$lib` o `$env` fallisce.
 
 - `intent` — le bande di intento d'acquisto e il loro ordine in coda. Zero dipendenze.
@@ -17,6 +17,9 @@ Ora quattro moduli stanno in `packages/leads-core/`, e il guardiano che già esi
   risponde alla stessa domanda di `platformOf`: dove si raggiunge questa persona.
 - `prompts` — il drafter del commento e del DM, con i guard-rail nati dai modi in cui le
   bozze fallivano davvero, e la selezione degli N migliori del giorno.
+- `feed` — dove si trovano le conversazioni: RSS, Google News, subreddit in rising, e le
+  ricerche su Reddit, Threads, X Communities e LinkedIn. Più il parsing, le finestre di
+  freschezza e il round-robin a quote eque.
 
 ## Le due dipendenze invertite, e le due lasciate stare
 
@@ -33,12 +36,30 @@ fuori anche `radarScan`, `radarProduce`, `radarArticles`, il digest e il tick: s
 orchestrazione, e tirarli dentro avrebbe voluto dire una `Deps` con quindici campi per
 guadagnare niente.
 
-## Cosa NON è stato fatto
+## I fetcher: un factory, non quindici parametri
 
-I fetcher delle sorgenti (Reddit, Threads, X Communities, LinkedIn, RSS) sono la parte più
-riusabile di tutte e sono ancora in `radar.ts`: dipendono da `scrapeCreatorsGet`, da quattro
-variabili d'ambiente e dal fallback Browserless. Sono il prossimo lotto, e sono ~350 righe:
-tenerlo separato lascia questo diff verificabile invece che enorme.
+I fetcher sono la parte più riusabile di tutte, e l'unica con dipendenze vere: il gateway di
+scraping, il token del feed personale di Reddit, e Chrome vero per i feed che bloccano i
+client HTTP da server. Non basta un `deps` per funzione, perché `radarScan` li passa come
+callback per le query dinamiche e `radarEngage` usa `fetchRedditText` da solo — quindi
+`createSources(deps)` restituisce le funzioni già legate, e i call site restano puliti.
+
+`redditAuth` è una **funzione** e non un oggetto: il codice originale leggeva `env` a ogni
+chiamata, e congelare le credenziali alla costruzione del factory avrebbe voluto dire che un
+cambio d'ambiente non arriva mai. C'è un test che lo pinna, perché è esattamente il tipo di
+regressione che nessuno nota finché non serve.
+
+Lo script di pagina di Browserless resta nell'app: il package riceve "una funzione che dato un
+URL torna il testo" e non sa che dietro c'è un browser.
+
+È stato aggiunto un test che non esisteva, sull'instradamento di ogni `kind` verso il suo
+endpoint. È la classe di bug che i commenti del file raccontano — `linkedin_query` che cadeva
+su `fetchFeed` e scaricava le parole chiave come se fossero un URL RSS, e X chiamata con `id=`
+dove l'endpoint vuole `url=`. Entrambi si presentavano come "0 item": un successo pulito e
+vuoto, indistinguibile da "nessuna conversazione degna oggi". Ora un'inversione di quel tipo
+fa fallire un test invece di spegnere una sorgente in silenzio.
+
+## Difetti visti e non toccati
 
 Due difetti trovati leggendo, non toccati per non mescolare spostamento e comportamento:
 `radar.ts` importava `suppressAuthor` senza chiamarlo mai (rimosso, era import morto), e il
